fixing int8 type and compile warnings

+++ source/mtree_int8.h
@@ -9,7 +9,7 @@
 #include "access/gist.h"
 #include "mtree_gist.h"
 
-//#define MTREE_INT8_SIZE (2 * sizeof(int) + sizeof(long long int)) // 16 bytes
+//#define MTREE_INT8_SIZE (2 * sizeof(int) + sizeof(long long)) // 16 bytes
 #define MTREE_INT8_SIZE sizeof(mtree_int8) // 16 bytes
 #define DatumGetMtreeInt8(x) ((mtree_int8 *) PG_DETOAST_DATUM(x))
 #define PG_GETARG_MTREE_INT8_P(x) DatumGetMtreeInt8(PG_GETARG_DATUM(x))
@@ -17,8 +17,8 @@
 
 typedef struct {
 	int parentDistance;
-	int coveringRadius;
-	long long int data;
+	long long coveringRadius;
+	long long data;
 } __attribute__((packed, aligned(1))) mtree_int8;
 
 #endif
